model/requirements-pip-gpu.txt
backports.weakref==1.0rc1
bleach==1.5.0
boto==2.48.0
bz2file==0.98
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
gensim==2.2.0
html5lib==0.9999999
idna==2.5
Markdown==2.6.8
natto-py==0.8.0
mojimoji==0.0.5
numpy==1.13.1
protobuf==3.3.0
pycparser==2.18
requests==2.18.1
scikit-learn==0.18.2
scipy==0.19.1
six==1.10.0
smart-open==1.5.3
tensorflow-gpu==1.2.1
urllib3==1.21.1
Werkzeug==0.12.2
